requirements-prod.txt
numpy>=1.26
onnx>=1.16
onnxscript>=0.1
onnxruntime>=1.18
scikit-learn>=1.4
torch>=2.2
torchvision>=0.17
timm>=1.0
open_clip_torch>=3.0
transformers>=4.40,<5
